# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fountainhead API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Submit a quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Quote submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "400":
          description: Invalid request
  /quotes/estimate:
    post:
      operationId: estimateQuote
      tags: [quotes]
      summary: Get price estimate for a quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateQuoteRequest"
      responses:
        "200":
          description: Price estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteEstimate"

  # --- TESTIMONIALS ---
...
components:
  schemas:
...
    QuoteServiceType:
      type: string
      enum: [mixing, mastering, stem-mastering, production, mixing-mastering, string-arrangement, cowriting, guitar-solo]
    EstimateQuoteRequest:
      type: object
      properties:
        serviceType:
          $ref: "#/components/schemas/QuoteServiceType"
        songCount:
          type: integer
          default: 1
        songsOver6Min:
          type: integer
          nullable: true
          description: Number of songs longer than 6 minutes (+10% each)
        additionalServices:
          type: array
          items:
            type: string
          nullable: true
      required: [serviceType, songCount]
    QuoteEstimate:
      type: object
      properties:
        serviceType:
          $ref: "#/components/schemas/QuoteServiceType"
        basePrice:
          type: number
        totalPrice:
          type: number
        songCount:
          type: integer
        breakdown:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              amount:
                type: number
            required: [label, amount]
      required: [serviceType, basePrice, totalPrice, songCount, breakdown]
    CreateQuoteRequest:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        serviceType:
          $ref: "#/components/schemas/QuoteServiceType"
        songCount:
          type: integer
          default: 1
        projectDescription:
          type: string
        genre:
          type: string
          nullable: true
        budget:
          type: number
          nullable: true
        referenceLinks:
          type: string
          nullable: true
        additionalServices:
          type: array
          items:
            type: string
          nullable: true
      required: [customerName, customerEmail, serviceType, songCount, projectDescription]
    Quote:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        serviceType:
          $ref: "#/components/schemas/QuoteServiceType"
        songCount:
          type: integer
        projectDescription:
          type: string
        genre:
          type: string
          nullable: true
        budget:
          type: number
          nullable: true
        estimatedPrice:
          type: number
        status:
          type: string
          enum: [pending, reviewed, accepted, declined]
        createdAt:
          type: string
      required: [id, customerName, customerEmail, serviceType, songCount, projectDescription, estimatedPrice, status, createdAt]

    # --- TESTIMONIALS ---
...